Human-Readable Matches

Converts a vecvec object into a human-readable format.

Usage

human_readable(.vecvec, match_vars = c("whole", "token", "substring", "fuzzy"))

Arguments

.vecvec

A vecvec object.

match_vars

A character vector of specifying which matches to return. Defaults to all matches:

whole: exact whole word matches with all spaces stripped out

token: exact whole token matches

substring: exact substring match

fuzzy: fuzzy text match using OSA - strict, max distance == 1

Value

A data frame of human-readable matches.

Examples

vecvec(create_random_names(), create_random_names()) |> human_readable()
#> # A tibble: 140 × 4
#>    original_row_id primary             lookup_id lookup_value
#>              <int> <chr>                   <int> <chr>       
#>  1               1 Charlotte                  79 charlotte   
#>  2               4 white                      73 white       
#>  3               5 Christopher Sanchez         4 christopher 
#>  4               5 Christopher Sanchez        60 SANCHEZ     
#>  5               5 Christopher Sanchez        12 Christopher 
#>  6               5 Christopher Sanchez        76 christopher 
#>  7               5 Christopher Sanchez        87 Sanchez     
#>  8               7 Sanchez                    60 SANCHEZ     
#>  9               7 Sanchez                    87 Sanchez     
#> 10               9 ASHER SCOTT                13 scott       
#> # ℹ 130 more rows